fix(stm32mp2): disable branch protection
STM32MP2 is based on Cortex-A35, wich is an Armv8.0. BTI is only
available starting Armv8.5. Although the added code with the feature is
replaced with NOP, it adds several lines of unused code, which grows
the size of TF-A binary. This code is removed with the compilation
option: -mbranch-protection=none.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I47e108906024196c794783bfd00a9d15301031f7
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/292071
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>