feat(stm32mp2): add supported state id in device tree
Allow board to reduce the list of supported state IDs for PSCI suspend
request, CPU or system.

This feature is useful for board without PMIC, when the supported
modes are limited by the regulators present on the board. For example
when the 2 LPLV modes are not supported as the associated regulators are
fixed.

To remove a mode, the associated sub-node in domain-idle-states node
should be removed from the board device tree.

Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Change-Id: I65c5df119ff289e4f57fdd69e29e5961440e8ad8
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/312839
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>