feat(st): introduce EARLY traces macros
In case STM32MP_EARLY_CONSOLE, the EARLY will just use the equivalent
tf_log macros. If not, the message will just be discarded. This will be
used when the full console driver is not initialized, during init phase
or in drivers called before console is set, like clock driver. This also
reduces the size of the binary when STM32MP_EARLY_CONSOLE is not set.

Change-Id: I5704c3ec658af4e32c69d310edd74206da04aaee
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/338367